Dynamically adapt the CCC based on SiStrip detector
The logic of the CCC has been modified so that, for TOB and TEC, the
cut selected for the specific iterations is dynamically lowered by 15%
(i.e. multiplied by 0.85). This in order to recover hits on track for
the detector in which the loss of efficiency was more pronounced. The
effects have been quickly investigated by computing the number of
tracks for which the stop-reason was 'max consecutive lost hits',
before and after this change.