Add the possibility to use CCC as a Trajectory Filter
The CCC has been implemented as a TrajectoryFilter, in order to give
more flexibility to the user in selecting possible bad (from the
Cluster Charge point of view) hits. The counter of the number of the
hits that would have been removed by the CCC if applied as an
estimator is left as a paramter, whose current values is 3 (i.e. we
tolerate up to 2 bad hits, after which we discard the trajectory). The
final trajectory, if discarded due to eccessive number of bad CCC
hits, will anyhow contain the would-have-been removed 2 hits, since
they passed all the other estimators. The user has the possibility to
disentangle lost hits from bad for CCC and apply different criteria
with different thresholds during Patter Recognition and at filtering
level. In order to assess the validity of this PR it would be better
to turn completely of the CCC at estimator level.