Moved the CCC selection as a TrajectoryFilter into the Trajectory
In the previous implementation, the counting of the hits that should
be considered as bad for CCC was done inside the filter itself, every
time looping on all TrajectoryMeasurements associated to the
Trajectory. With this improved implentation the counting is done
directly while propagating the Trajectory. The number of bad hits for
CCC is locally cached in the {Temp,}Trajectory and updated only with
the measurements of the appended segments.