Protect access to updatedState
While testing if a newly added TrajectoryMeasurement to a Trajectory is
compatible or not with the cut on the cluster charge, a protection has
been added to test the validity of the updatedState before using it.
This fix a segmentation fault that was due to an invalid updatedState
caused by a failed KfUpdator. Full log avaialable here:
https://hypernews.cern.ch/HyperNews/CMS/get/recoDevelopment/1436.html

See also Pull Request #14060 for the corresponding PR for the 80X
branch.