Enable OfflinePV and Beamspot in trackingLowPU era
The trackingLowPU era has been cooked to perform tracking
also in the B=0 case, when the Pixel detector could be
excluded from data taking. Tracking was already working, but
the configuration for primary vertices and beamspot
calculation had to be tuned to properly work also in the
case in which Pixels are out of data taking. The code has
been successfully tested on Run 269207.