Bring back stuff from global to local memory
The timing of the heaviest function in the Riemann fit went
up to 3.3ms after we moved most of the local matrices to
global memory. Now we bring some of them back, while keeping
the number of used registers to a reasonable level. This
brings back the timing to around 2ms and still allows us to
run many kernel in parallel.