Riemann Fit MS Reworked
The Riemann Fit has been reworked so that both barrel and
forward cases are naturally supported without branching. The
underlying assumption is the uniform material distribution
within the Pixel Tracker. Also the line fit has been
reworked and is now using an ordinary least square fit in
the S-Z plane. The motivations and explanation are inside
the comments into the code.