Mate HGCalValidator MT safe
The HGCalValidator is a Global Module and, as such, can be called by
different threads, concurrently. In the previous implementation, some
event-specific map was build by an helper class (HGVHistoProducerAlgo)
on the heap. That map was then filled concurrently by several threads,
causing segmentation fault. The behaviour has been modified moving the
map on the "stack of HGCalValidator", thou making it local to each
thread invocation, and percolating it down to the helper class. Tested
with up to 24 streams with 24 threads. Results identical 1-to-1 if
compared against the single thread running output.