Remove unused label parameter
Actually, besides the fact that the label parameter is no longer used,
adding it as a python attribute to any module is rather dangerous, since
it will completely hide the label() method that is usually internally
handled by the cms.XYZ python classes. There are few other cases in
CMSSW where label is used as is as a configuration parameter: that
should be prevented, imho.