Move Trackster from struct to class
This move required quite a bit of editing all over several places.
Methods that could be gathered have been gathered. Direct access to
std containers has been granted, especially to keep the possibility to
use, e.g., back_inserter and similar utilities.
A dependency on Eigen has been added to the Trackster dataformats in
order to easily accommodate the handling of eigenvectors and
eigenvalues.
The overall final product contains a lot of boilerplate with respect to
the naive struct implementation. Yet to be carefully evaluated if it was
worth the effort.