Introduce penalty term in double alignment
When computing the alignment requirements between doublets, a penalty
term has been introduced to take into account the fact that clusters
with a small number of cells have their position determined with less
accuracy. The penalty term is such that it increments the compatibility
window by an additional 10% when the doublets involved are composes by a
total of 3 cells, and saturates when the number of cells is greater or
equal to 9. The parameters have been studied and tuned using a
singleMuon sample produced without noise and with a relaxed kappa
parameter.