Transverse and longitudinal critical distances
In the computation of the nearest cluster with higher local density, the
distance, in cm,  along the Z direction is also taken into account. This
should follow more closely the real 'energy-flow' within real showers,
avoid the ping-pong case of connections going back and forth while
slowly approaching the shower axis.

The longitudinal separation, express in number of layers, is also taken
into account while promoting a layercluster as a seed. The requirement
is in 'or' with the traditional one on the transverse distance. A
layercluster now, in order to be promoted as a seed, needs to be
separated at least threshold_xy distance on the transverse plane from its
nearest-higher OR separated at least threshold_layers from it. This will
avoid excessive fragmentation of showers along the Z direction due to
'energy fluctuations' in the overall 'energy flow'.

A new variable is now considered to promote a layer cluster as a seed.
It represent the ratio between the energy of the layer cluster and the
local density associated to it. A reasonable value for this variable
should be set to 1/(densitySiblingLayers+1), assuming a somewhat uniform
local density distribution.

Minor improvements in the printouts.