Add layerCluster producer
If TICLFromReco is run on top of the optimized HGCAL-only event content,
the producer is not needed. The local density map computed while
creating the layer clusters is part of the special HGCAL-only event
content. Unfortunately, that product is not persisted in any ordinary
data-tiers but is consumed by the HGCAL validation. Adding the
layercluster producer fixes the bug and makes the customize even more
generic.