Fix phi bin wrap
The method `searchBoxEtaPhi` of the Tile data structure returns the
correct bin numbers in eta and phi, but ignores the possible wrapping
problems that could happen when the phi range is crossing the
"bin-boundary". Therefore, a trivial loop on the client side between the
returned eta and phi bin ranges could result in less cells explored. The
solution implemented in `LinkingAlgoByPCAGeometric` was worst than the
original problem, since it caused a loop over the complementary range in
phi. The fix implemented here is temporary and is meant to be propagated
upstream inside `searchBoxEtaPhi`, with the corresponding changes in the
client code to perform a proper phi-wrapping before looping.