Add recoverBrokenHistory to CaloTruthAccumulator.
Adds a new option: `recoverBrokenHistory`.

When enabled, this option attempts to recover the history of vertices
that:
  - Have no associated parent `SimTrack`.
  - Do not originate from primary particles.

Without proper handling, these "orphan" vertices can lead to the
creation of separate `SimClusters` that are not associated with any
`CaloParticles`. his breaks the intended design and integrity of the
`CaloParticles``SimClusters` relationship.

The recovery mechanism works by assigning each orphan vertex to the
primary `SimTrack` associated with the `SimTracks` connected to that
orphan vertex. This recovered association is stored within each
`SimTrack`, "preserving" the consistency of the simulation history.