Bug fix for vertex splitting.
The call to cms::alpakatools::uniform_elements inside a
cms::alpakatools::independent_groups loop does not perform the full loop
on the requirested argument. Therefore the vertex splitting logic was
not applied to all vertices, but only to the one in the first block, at
most.
This fix will run the splitting logic on all vertices.
The verbose moda has been also extended in order to be able to identify
the bug in the first place and validate the fix, later on.